chore(stageTemplate): resolve comment with stage template
