+++ web/src/components/stageTemplate/add/template.module.less
@@ -1,7 +1,46 @@
+@import "../../../less/inc.var.less";
 .kv-item {
   :global {
     .item-name {
       min-width: 100px;
+    }
+  }
+}
+
+.stagetemplate-form {
+  :global {
+    .u-scm {
+      .u-scm-sellection {
+        margin-bottom: -25px;
+        .ant-radio-button-wrapper {
+          box-shadow: none !important;
+          margin-right: 16px;
+          border-radius: 4px;
+          border: 1px solid @color-border;
+          border-left: 1px solid @color-border;
+          &:not(:first-child)::before {
+            display: none;
+          }
+          &:first-child {
+            border-radius: 4px;
+          }
+          &:last-of-type {
+            margin-right: 0;
+            border-radius: 4px;
+          }
+        }
+        .ant-radio-button-wrapper-checked {
+          border-left: 1px solid @color-primary;
+          border: 1px solid @color-primary;
+        }
+      }
+    }
+    .loading {
+      text-align: center;
+      margin: 40px 0;
+    }
+    .tip-error {
+      color: red;
     }
   }
 }
