--- web/src/less/u.stageTemplate.less
@@ -1,35 +1,0 @@
-.stagetemplate-form {
-  .u-scm {
-    .u-scm-sellection {
-      margin-bottom: -25px;
-      .ant-radio-button-wrapper {
-        box-shadow: none !important;
-        margin-right: 16px;
-        border-radius: 4px;
-        border: 1px solid @color-border;
-        border-left: 1px solid @color-border;
-        &:not(:first-child)::before {
-          display: none;
-        }
-        &:first-child {
-          border-radius: 4px;
-        }
-        &:last-of-type {
-          margin-right: 0;
-          border-radius: 4px;
-        }
-      }
-      .ant-radio-button-wrapper-checked {
-        border-left: 1px solid @color-primary;
-        border: 1px solid @color-primary;
-      }
-    }
-  }
-  .loading {
-    text-align: center;
-    margin: 40px 0;
-  }
-  .tip-error {
-    color: red;
-  }
-}
